t/this-completion
This patch makes HaXe complete as if this. was typed before the cursor in this context:

class A {
  function x(){
    [ maybe command ;]
    CURSOR
  }

implementation details: HaXe is told there was a this. even if there isn't.
When completing in static function you either get an error
(this. can't be used in static context, eg trace(CURSOR )
or the list is empty.

Signed-off-by: Marc Weber <marco-oweber@gmx.de>